# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CollBar — Collective Bargaining Compensation Modeling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /salary-schedules:
    get:
      operationId: listSalarySchedules
      tags: [salary-schedules]
      summary: List salary schedules
      parameters:
        - name: bargainingUnitId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of salary schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalarySchedule"
    post:
      operationId: createSalarySchedule
      tags: [salary-schedules]
      summary: Create a salary schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalaryScheduleRequest"
      responses:
        "201":
          description: Created schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalarySchedule"
  /salary-schedules/{id}:
    get:
      operationId: getSalarySchedule
      tags: [salary-schedules]
      summary: Get a salary schedule with full grid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Salary schedule with lanes, steps, cells
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalaryScheduleWithGrid"
    delete:
      operationId: deleteSalarySchedule
      tags: [salary-schedules]
      summary: Delete a salary schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /hourly-schedules:
    get:
      operationId: listHourlySchedules
      tags: [hourly-schedules]
      summary: List hourly schedules
      parameters:
        - name: bargainingUnitId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of hourly schedules with categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourlyScheduleWithCategories"
    post:
      operationId: createHourlySchedule
      tags: [hourly-schedules]
      summary: Create a hourly schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHourlyScheduleRequest"
      responses:
        "201":
          description: Created hourly schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HourlyScheduleWithCategories"
...
  /heatmap/{scenarioId}:
    get:
      operationId: getHeatmapData
      tags: [heatmap]
      summary: Get heatmap data for a scenario (all years)
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
        - name: bargainingUnitId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Heatmap data for all contract years
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeatmapData"
...
components:
  schemas:
...
    SalarySchedule:
      type: object
      properties:
        id:
          type: string
        bargainingUnitId:
          type: string
        name:
          type: string
        effectiveYear:
          type: number
        baseSalary:
          type: string
        createdAt:
          type: string
      required:
        - id
        - bargainingUnitId
        - name
        - effectiveYear
        - baseSalary
        - createdAt
    Lane:
      type: object
      properties:
        id:
          type: string
        salaryScheduleId:
          type: string
        name:
          type: string
        displayOrder:
          type: number
        indexMultiplier:
          type: string
      required:
        - id
        - name
        - displayOrder
        - indexMultiplier
    Step:
      type: object
      properties:
        id:
          type: string
        salaryScheduleId:
          type: string
        stepNumber:
          type: number
        incrementMultiplier:
          type: string
      required:
        - id
        - stepNumber
        - incrementMultiplier
    ScheduleCell:
      type: object
      properties:
        id:
          type: string
        laneId:
          type: string
        stepId:
          type: string
        stepNumber:
          type: number
        salaryAmount:
          type: string
      required:
        - id
        - laneId
        - stepId
        - stepNumber
        - salaryAmount
    SalaryScheduleWithGrid:
      allOf:
        - $ref: "#/components/schemas/SalarySchedule"
        - type: object
          properties:
            lanes:
              type: array
              items:
                $ref: "#/components/schemas/Lane"
            steps:
              type: array
              items:
                $ref: "#/components/schemas/Step"
            cells:
              type: array
              items:
                $ref: "#/components/schemas/ScheduleCell"
    CreateSalaryScheduleRequest:
      type: object
      properties:
        bargainingUnitId:
          type: string
        name:
          type: string
        effectiveYear:
          type: number
        baseSalary:
          type: string
        lanes:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              displayOrder:
                type: number
              indexMultiplier:
                type: string
        steps:
          type: array
          items:
            type: object
            properties:
              stepNumber:
                type: number
              incrementMultiplier:
                type: string
        cells:
          type: array
          items:
            type: object
            properties:
              laneIndex:
                type: number
              stepNumber:
                type: number
              salaryAmount:
                type: string
      required:
        - bargainingUnitId
        - name
    HourlyCategory:
      type: object
      properties:
        id:
          type: string
        hourlyScheduleId:
          type: string
        name:
          type: string
        baseHourlyRate:
          type: string
        annualHours:
          type: string
        displayOrder:
          type: number
      required:
        - id
        - name
        - baseHourlyRate
        - annualHours
    HourlyScheduleWithCategories:
      type: object
      properties:
        id:
          type: string
        bargainingUnitId:
          type: string
        effectiveYear:
          type: number
        categories:
          type: array
          items:
            $ref: "#/components/schemas/HourlyCategory"
        createdAt:
          type: string
      required:
        - id
        - bargainingUnitId
        - effectiveYear
        - categories
        - createdAt
    CreateHourlyScheduleRequest:
      type: object
      properties:
        bargainingUnitId:
          type: string
        effectiveYear:
          type: number
        categories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              baseHourlyRate:
                type: string
              annualHours:
                type: string
              displayOrder:
                type: number
      required:
        - bargainingUnitId
...
    HeatmapYearData:
      type: object
      properties:
        contractYear:
          type: number
        yearLabel:
          type: string
        cells:
          type: array
          items:
            type: object
            properties:
              laneId:
                type: string
              laneName:
                type: string
              stepNumber:
                type: number
              employeeCount:
                type: number
              totalSalary:
                type: string
              employees:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    salary:
                      type: string
        lanes:
          type: array
          items:
            $ref: "#/components/schemas/Lane"
        maxStep:
          type: number
        totalEmployees:
          type: number
        medianSalary:
          type: string
          nullable: true
        avgStep:
          type: number
          nullable: true
        employeesAtTopStep:
          type: number
      required:
        - contractYear
        - yearLabel
        - cells
        - lanes
        - maxStep
        - totalEmployees
        - employeesAtTopStep
    HeatmapData:
      type: object
      properties:
        scenarioId:
          type: string
        bargainingUnitId:
          type: string
        bargainingUnitName:
          type: string
        years:
          type: array
          items:
            $ref: "#/components/schemas/HeatmapYearData"
      required:
        - scenarioId
        - bargainingUnitId
        - years
...