# lib/api-zod/src/generated/types/createSalaryScheduleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CollBar — Collective Bargaining Compensation Modeling API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSalaryScheduleRequestCellsItem } from "./createSalaryScheduleRequestCellsItem";
import type { CreateSalaryScheduleRequestLanesItem } from "./createSalaryScheduleRequestLanesItem";
import type { CreateSalaryScheduleRequestStepsItem } from "./createSalaryScheduleRequestStepsItem";

export interface CreateSalaryScheduleRequest {
  bargainingUnitId: string;
  name: string;
  effectiveYear?: number;
  baseSalary?: string;
  lanes?: CreateSalaryScheduleRequestLanesItem[];
  steps?: CreateSalaryScheduleRequestStepsItem[];
  cells?: CreateSalaryScheduleRequestCellsItem[];
}