# lib/api-zod/src/generated/types/scheduleCell.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CollBar — Collective Bargaining Compensation Modeling API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleCell {
  id: string;
  laneId: string;
  stepId: string;
  stepNumber: number;
  salaryAmount: string;
}