# lib/api-zod/src/generated/types/step.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CollBar — Collective Bargaining Compensation Modeling API
 * OpenAPI spec version: 0.1.0
 */

export interface Step {
  id: string;
  salaryScheduleId?: string;
  stepNumber: number;
  incrementMultiplier: string;
}